Improve the search bar on the managers page
This commit changes the search field on the managers page from
an input box to a `Search` widget, which improves the styling and
wires up the search functions.

We also convert the Managers object to a `Component` to prepare
for wiring it with real data.  It still uses the hard-coded
data for now, but the search box can successfully filter against
the hard-coded data.